Fix d2l.page.load event not being measured in IE11
