+++ bsi.html
@@ -1,4 +1,4 @@
-<script src="src/page-loading/page-loading.js"></script>
+<script src="src/page-loading/web-components-ready.js"></script>
 
 <link rel="import" href="bsi-style.html">
 <link rel="import" href="../d2l-alert/d2l-alert.html">
